Add files via upload
First, I has added an Enumerable class call Operation, this is used in each class that extends from LogicalOperation in order to have in those a variable call operation which is the value at that spot (it is also what is printed in the toString()). negateAll() returns a list of components actually modified with the negations meanwhile negate() returns a LogicalOperation(). negateAll() creates an auxiliar list<LogicalOperations> and fill that one with the negation of each of the branches of components.

So, the main idea of the program in my head at least to the point of signar is: 
          There will be a functional (xD)  method call next() which traverse the tree untill it finds a notOperation when it does it, it will do a           negateAll() to its components. This will may eventually get to the method negate of the notOperation which depending on things will start the negateAll() or stop it. There are some coments about this in the code. Hope you like my bloodborne dlc i am gonna watch a streamer of it instead.... T.T :PP