New Operation, Literal and Element
Now, both Operation and Literal are considered Elements (implements the
interface). Also, the NOT operation have been removed as Signar will put
all NOT operations on literals, so only them have to be in charge of
that. I've also added the method addComponent(element) on the Operation
class and removed the components from the constructor, so now we build
the semantic tree from top to bottom. Finally, I've removed the files
AndOperation, OrOperation and NotOperation and upgrated the toString
method of the Operation and Literal class.